[Fix #124] room폴더 구조 변경
