+++ api/liveKitApi.ts
@@ -1,9 +1,11 @@
-import { MafiaRoom } from "../types";
+import { getUserInfo } from "@/utils/supabase/authAPI";
 
-export async function getToken({ room, userInfo }: MafiaRoom) {
+export async function getToken(room: string) {
+  const userInfo = await getUserInfo();
+
   const response = await fetch(
     `/livekit/get-participant-token?room=${room}&userId=${userInfo!.id}&userNickname=${
-      userInfo!.user_metadata.userNickname
+      userInfo!.user_metadata.nickname
     }`,
     {
       cache: "no-store"
