+++ app/livekit/get-participant-token/route.ts
@@ -21,7 +21,7 @@ export async function GET(req: NextRequest) {
     return NextResponse.json({ error: "Server misconfigured" }, { status: 500 });
   }
 
-  const at = new AccessToken(apiKey, apiSecret, { identity: userId });
+  const at = new AccessToken(apiKey, apiSecret, { identity: nickname });
 
   at.addGrant({ room, roomJoin: true, canPublish: true, canSubscribe: true });
